refactor: breakup main functions into separate files
to prevent circular imports and make the code easier to maintain
